code/requirements.txt
pystray
Pillow
pytesseract
pyserial
keyboard
pyperclip
tk
